Features/auth

+++ components/client/CCAllVolunteers.tsx
@@ -1,14 +1,16 @@
 "use client";
 
 import { useRouter } from "next/navigation";
-import { Fragment, useMemo, useState } from "react";
+import { useMemo, useState } from "react";
 import DataTable from "react-data-table-component";
 import { PiLegoSmiley, PiLegoSmileyDuotone } from "react-icons/pi";
-import CpInput from "./Input";
+import GCInputSearch from "@/components/global/GCInputSearch";
 import { IoPersonAdd } from "react-icons/io5";
 import { Tooltip } from "react-tooltip";
 import { RiDeleteBinLine } from "react-icons/ri";
 import { deleteVolunteer } from "@/utils/apis/delete";
+import { toast } from "react-toastify";
+import { sleep } from "@/utils/helpers";
 
 interface Data {
   _id: string
@@ -19,6 +21,13 @@ interface Data {
   status: string
   segment: string
   active: boolean
+}
+
+interface Columns {
+  name: string
+  selector: (row: Data) => string
+  sortable: boolean
+  button?: boolean
 }
 
 const conditionalRowStyles = [
@@ -37,23 +46,28 @@ const conditionalRowStyles = [
   }
 ];
 
-export default function CpAllVolunteers({data}: {data: Data[]}) {
+export default function CCAllVolunteers({ data, isAdmin }: { data: Data[], isAdmin: boolean }) {
   const router = useRouter();
   const [query, setQuery] = useState('');
-  const isAdmin = true;
 
   const deleteVol = async (volunteerId: string, volunteerName: string) => {
     const confirmed = confirm(`Are you sure you want to remove ${volunteerName} as volunteer?`);
     if (!confirmed) return;
 
-    const res = await deleteVolunteer(volunteerId);
-    if (res.success) {
-      alert(res.message);
+    try {
+      const res = await deleteVolunteer(volunteerId);
+      if (res.success) {
+        toast.info(res.message);
+      }
+    } catch (error) {
+      toast.error("Failed to remove volunteer!");
+    } finally {
+      await sleep(1000);
       router.refresh();
     }
   }
 
-  const columns = [
+  const columns: Columns[] = [
     {
       name: "First Name",
       selector: (row: Data) => row.firstName,
@@ -79,13 +93,16 @@ export default function CpAllVolunteers({data}: {data: Data[]}) {
       selector: (row: Data) => row.status,
       sortable: true,
     },
-    {
+  ];
+
+  if (isAdmin) {
+    columns.push({
       name: "Actions",
       selector: (row: Data) => <RiDeleteBinLine onClick={() => deleteVol(row._id, row.name)} size={18} /> as any,
       sortable: false,
       button: true,
-    }
-  ];
+    })
+  }
 
   const filteredVolunteers = useMemo(() => {
     const filteredValues = data.filter((volunteer) => {
@@ -109,7 +126,8 @@ export default function CpAllVolunteers({data}: {data: Data[]}) {
   }
 
   return (
-    <Fragment>
+    <div className="flex justify-center">
+      <div className="w-full">
         <div className="flex flex-col gap-7 text-slate-700 px-32">
           <div className="flex justify-between">
             <h1 className="text-2xl">
@@ -117,18 +135,20 @@ export default function CpAllVolunteers({data}: {data: Data[]}) {
             </h1>
             <div className="flex gap-3 justify-end w-1/2">
               <div className="w-1/2">
-                <CpInput onChange={(event) => setQuery(event.target.value)} />
+                <GCInputSearch onChange={(event) => setQuery(event.target.value)} />
               </div>
-              <button
-                onClick={openModal}
-                id="add-volunteer"
-                className="flex flex-col justify-center bg-sky-600 py-1 px-2 rounded-xl h-10"
-              >
-                <IoPersonAdd size={24} className="text-slate-50" />
-              </button>
+              { isAdmin &&
+                <button
+                  onClick={openModal}
+                  id="add-volunteer"
+                  className="flex flex-col justify-center bg-sky-600 py-1 px-2 rounded-xl h-10"
+                >
+                  <IoPersonAdd size={24} className="text-slate-50" />
+                </button>
+              }
             </div>
           </div>
-          <div className="py-1 border border-slate-200 rounded-md">
+          <div className={`${isAdmin && !!filteredVolunteers.length ? "border border-slate-200" : ""} py-1 rounded-md`}>
             <DataTable
               columns={columns}
               data={filteredVolunteers}
@@ -155,6 +175,7 @@ export default function CpAllVolunteers({data}: {data: Data[]}) {
           </div>
         </div>
         <Tooltip variant="info" anchorSelect="#add-volunteer" place="top-end">Add Volunteer</Tooltip>
-    </Fragment>
+      </div>
+    </div>
   )
 }
